requirements.txt
flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
textblob>=0.18.0
joblib>=1.3.0
Pillow>=10.0.0
gunicorn>=21.2.0
